// sourcecode/BTrak.Packages/snova-trading-module/trading-module-components/trading-components/src/lib/billing/models/template-config-model.ts
export class TemplateConfigModel {
    templateConfigurationName: string;
    templateConfiguration: string;
    templateConfigurationId: string;
    templateTypeName: string;
    templateTypeId: string;
    isArchived: boolean;
    timeStamp: any;
    legalEntityId: string;
    clientTypeId: string;
}